Current Dog and Guard Behavior
Committing and pushing what I have so far with the guard and dog behavior. Taking a break for my right eye. Will be finishing refinement of Dog Chase and LOS tomorrow, adjusting Handler following and merging Guard and Handler into one file. Will be adjusting Dog melee attack this weekend and expanding on the turrets.